backend/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
PyJWT==2.0.1
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
